backend/requirements.txt
fastapi==0.104.1
uvicorn==0.23.2
python-multipart==0.0.6
pypdf==3.17.0
pillow==10.1.0
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.196
pydantic==2.4.2

